+++ examples/gnn_node.py
@@ -7,23 +7,19 @@ from typing import Dict, List
 import numpy as np
 import torch
 from inferred_stypes import dataset2inferred_stypes
+from model import Model
 from text_embedder import GloveTextEmbedding
-from torch import Tensor
 from torch.nn import BCEWithLogitsLoss, L1Loss
 from torch_frame.config.text_embedder import TextEmbedderConfig
-from torch_frame.data import TensorFrame
 from torch_geometric.data import HeteroData
 from torch_geometric.loader import NeighborLoader
-from torch_geometric.nn import MLP
 from torch_geometric.seed import seed_everything
-from torch_geometric.typing import EdgeType, NodeType
 from tqdm import tqdm
 
-from relbench.data import RelBenchDataset
+from relbench.data import NodeTask, RelBenchDataset
 from relbench.data.task_base import TaskType
 from relbench.datasets import get_dataset
 from relbench.external.graph import get_node_train_table_input, make_pkey_fkey_graph
-from relbench.external.nn import HeteroEncoder, HeteroGraphSAGE, HeteroTemporalEncoder
 
 parser = argparse.ArgumentParser()
 parser.add_argument("--dataset", type=str, default="rel-stackex")
@@ -45,7 +41,7 @@ root_dir = "./data"
 
 # TODO: remove process=True once correct data/task is uploaded.
 dataset: RelBenchDataset = get_dataset(name=args.dataset, process=True)
-task = dataset.get_task(args.task, process=True)
+task: NodeTask = dataset.get_task(args.task, process=True)
 
 col_to_stype_dict = dataset2inferred_stypes[args.dataset]
 
@@ -81,7 +77,6 @@ for split, table in [
         persistent_workers=args.num_workers > 0,
     )
 
-
 clamp_min, clamp_max = None, None
 if task.task_type == TaskType.BINARY_CLASSIFICATION:
     out_channels = 1
@@ -98,69 +93,14 @@ elif task.task_type == TaskType.REGRESSION:
         task.train_table.df[task.target_col].to_numpy(), [2, 98]
     )
 
-
-class Model(torch.nn.Module):
-    def __init__(self):
-        super().__init__()
-
-        self.encoder = HeteroEncoder(
-            channels=args.channels,
-            node_to_col_names_dict={
-                node_type: data[node_type].tf.col_names_dict
-                for node_type in data.node_types
-            },
-            node_to_col_stats=data.col_stats_dict,
-        )
-        self.temporal_encoder = HeteroTemporalEncoder(
-            node_types=[
-                node_type for node_type in data.node_types if "time" in data[node_type]
-            ],
-            channels=args.channels,
-        )
-        self.gnn = HeteroGraphSAGE(
-            node_types=data.node_types,
-            edge_types=data.edge_types,
-            channels=args.channels,
-            aggr=args.aggr,
-            num_layers=args.num_layers,
-        )
-        self.head = MLP(
-            args.channels,
-            out_channels=out_channels,
-            norm="batch_norm",
-            num_layers=1,
-        )
-
-    def forward(
-        self,
-        tf_dict: Dict[NodeType, TensorFrame],
-        edge_index_dict: Dict[EdgeType, Tensor],
-        seed_time: Tensor,
-        time_dict: Dict[NodeType, Tensor],
-        batch_dict: Dict[NodeType, Tensor],
-        num_sampled_nodes_dict: Dict[NodeType, List[int]],
-        num_sampled_edges_dict: Dict[EdgeType, List[int]],
-    ) -> Tensor:
-        x_dict = self.encoder(tf_dict)
-
-        rel_time_dict = self.temporal_encoder(seed_time, time_dict, batch_dict)
-        for node_type, rel_time in rel_time_dict.items():
-            x_dict[node_type] = x_dict[node_type] + rel_time
-
-        x_dict = self.gnn(
-            x_dict,
-            edge_index_dict,
-            num_sampled_nodes_dict,
-            num_sampled_edges_dict,
-        )
-
-        out = self.head(x_dict[entity_table][: seed_time.size(0)])
-        if not self.training and task.task_type == TaskType.REGRESSION:
-            out = torch.clamp(out, clamp_min, clamp_max)
-        return out
-
-
-model = Model().to(device)
+model = Model(
+    data=data,
+    num_layers=args.num_layers,
+    channels=args.channels,
+    out_channels=out_channels,
+    aggr=args.aggr,
+    norm="batch_norm",
+).to(device)
 optimizer = torch.optim.Adam(model.parameters(), lr=args.lr)
 
 
@@ -173,6 +113,7 @@ def train() -> Dict[str, float]:
 
         optimizer.zero_grad()
         pred = model(
+            task.entity_table,
             batch.tf_dict,
             batch.edge_index_dict,
             batch[entity_table].seed_time,
@@ -200,6 +141,7 @@ def test(loader: NeighborLoader) -> np.ndarray:
     for batch in tqdm(loader):
         batch = batch.to(device)
         pred = model(
+            task.entity_table,
             batch.tf_dict,
             batch.edge_index_dict,
             batch[entity_table].seed_time,
@@ -207,6 +149,8 @@ def test(loader: NeighborLoader) -> np.ndarray:
             batch.batch_dict,
             batch.num_sampled_nodes_dict,
             batch.num_sampled_edges_dict,
+            clamp_min=clamp_min,
+            clamp_max=clamp_max,
         )
 
         if task.task_type == TaskType.BINARY_CLASSIFICATION:
